Update src/CMakeLists.txt
This correctly sets the rpath prefix for libcyclopts, which then ensures it is correctly found by libcycluscore.dylib and the cyclus executable on OSX-based systems. This accounts for the different dynamic library loading behavior unique to OSX.

Also replacing "dl" with ${CMAKE_DL_LIBS} in the TARGET_LINK_LIBRARIES command, as this is more OS-general.